backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
python-dotenv
google-generativeai
